Move global (load/save) methods into module, etc
A whole bunch of incremental changes as we experiment with creating the
command line utility. At this point we're reading and writing to the
store when we need to.

- Install lodash (with types)
- Change the anchor command to just `add` and include basic
functionality
- Change the go command so that it now properly looks for data that
might come from the store
- Tighten up a couple of type definitions, including some around the
mutability of the store (that will help us avoid certain mistakes)
- Move a bunch of the global logic around loading and saving the store
into the global folder.
- Begin defining some patterns that we'll use for commands, mainly that
you have getter functions to pull in the store or config if needed, and
there is a method to update the store in responds as well.